Deduplicate mock AsyncCamera
(Assisted by Claude Code; any errors are mine.)
